[chores] Moved logic to device sub app
