+++ openwisp_monitoring/device/apps.py
@@ -1,12 +1,14 @@
+from copy import deepcopy
 from urllib.parse import urljoin
 
 from django.apps import AppConfig
 from django.conf import settings
 from django.core.cache import cache
+from django.db.models import Case, Count, Sum, When
 from django.db.models.signals import post_delete, post_save
 from django.urls import reverse_lazy
 from django.utils.translation import gettext_lazy as _
-from swapper import load_model
+from swapper import get_model_name, load_model
 
 from openwisp_controller.config.signals import checksum_requested, config_status_changed
 from openwisp_controller.connection import settings as connection_settings
@@ -16,6 +18,7 @@ from openwisp_utils.admin_theme import (
     register_dashboard_chart,
     register_dashboard_template,
 )
+from openwisp_utils.admin_theme.menu import register_menu_group
 
 from ..check import settings as check_settings
 from ..settings import MONITORING_API_BASEURL, MONITORING_API_URLCONF
@@ -38,6 +41,7 @@ class DeviceMonitoringConfig(AppConfig):
         self.device_recovery_detection()
         self.set_update_config_model()
         self.register_dashboard_items()
+        self.register_menu_groups()
         self.add_connection_ignore_notification_reasons()
 
     def connect_device_signals(self):
@@ -243,6 +247,69 @@ class DeviceMonitoringConfig(AppConfig):
                     'monitoring_location_geojson_url': loc_geojson_url,
                 },
             )
+
+        if app_settings.WIFI_SESSIONS_ENABLED:
+            WifiSession = load_model('device_monitoring', 'WifiSession')
+            register_dashboard_chart(
+                position=6,
+                config={
+                    'name': _('Currently Active WiFi Sessions'),
+                    'query_params': {
+                        'app_label': WifiSession._meta.app_label,
+                        'model': WifiSession._meta.model_name,
+                        'annotate': {
+                            'active': Count(
+                                Case(
+                                    When(
+                                        stop_time__isnull=True,
+                                        then=1,
+                                    )
+                                )
+                            ),
+                        },
+                        'aggregate': {
+                            'active__sum': Sum('active'),
+                        },
+                    },
+                    'filters': {
+                        'key': 'stop_time__isnull',
+                        'active__sum': 'true',
+                    },
+                    'colors': {
+                        'active__sum': '#267126',
+                    },
+                    'labels': {
+                        'active__sum': _('Currently Active WiFi Sessions'),
+                    },
+                    'quick_link': {
+                        'url': reverse_lazy(
+                            'admin:{app_label}_{model_name}_changelist'.format(
+                                app_label=WifiSession._meta.app_label,
+                                model_name=WifiSession._meta.model_name,
+                            )
+                        ),
+                        'label': _('Open WiFi session list'),
+                    },
+                },
+            )
+
+    def register_menu_groups(self):
+        from ..monitoring.apps import MonitoringConfig
+
+        menu_config = deepcopy(MonitoringConfig.menu_group_config)
+        if app_settings.WIFI_SESSIONS_ENABLED:
+            menu_config['items'].update(
+                {
+                    1: {
+                        'label': _('WiFi Sessions'),
+                        'model': get_model_name('device_monitoring', 'WifiSession'),
+                        'name': 'changelist',
+                        'icon': 'ow-monitoring-checks',
+                    },
+                }
+            )
+        register_menu_group(position=80, config=menu_config)
+        del MonitoringConfig
 
     def add_connection_ignore_notification_reasons(self):
         ConnectionConfig._ignore_connection_notification_reasons.extend(
