+++ openwisp_monitoring/device/settings.py
@@ -50,3 +50,4 @@ AUTO_CLEAR_MANAGEMENT_IP = get_settings_value('AUTO_CLEAR_MANAGEMENT_IP', True)
 DEVICE_RECOVERY_DETECTION = get_settings_value('DEVICE_RECOVERY_DETECTION', True)
 MAC_VENDOR_DETECTION = get_settings_value('MAC_VENDOR_DETECTION', True)
 DASHBOARD_MAP = get_settings_value('DASHBOARD_MAP', True)
+WIFI_SESSIONS_ENABLED = get_settings_value('WIFI_SESSIONS_ENABLED', True)
