+++ openwisp_monitoring/monitoring/tasks.py
@@ -1,6 +1,5 @@
 from celery import shared_task
 from django.core.exceptions import ObjectDoesNotExist
-from django.utils.timezone import now, timedelta
 from swapper import load_model
 
 from ..db import timeseries_db
@@ -51,60 +50,3 @@ def migrate_timeseries_database():
     )
 
     migrate_influxdb_structure()
-
-
-@shared_task
-def save_wifi_clients_and_sessions(device_data, device_pk):
-    _WIFICLIENT_FIELDS = ['vendor', 'ht', 'vht', 'wmm', 'wds', 'wps']
-    WifiClient = load_model('monitoring', 'WifiClient')
-    WifiSession = load_model('monitoring', 'WifiSession')
-
-    active_sessions = []
-    interfaces = device_data.get('interfaces', [])
-    for interface in interfaces:
-        if interface.get('type') != 'wireless':
-            continue
-        interface_name = interface.get('name')
-        wireless = interface.get('wireless', {})
-
-        ssid = wireless.get('ssid')
-        clients = wireless.get('clients', [])
-        for client in clients:
-            # Save WifiClient
-            client_obj, created = WifiClient.objects.get_or_create(
-                mac_address=client.get('mac')
-            )
-            update_fields = []
-            for field in _WIFICLIENT_FIELDS:
-                if getattr(client_obj, field) != client.get(field):
-                    setattr(client_obj, field, client.get(field))
-                    update_fields.append(field)
-            if update_fields:
-                client_obj.full_clean()
-                client_obj.save(update_fields=update_fields)
-
-            # Save WifiSession
-            session_obj, _ = WifiSession.objects.get_or_create(
-                device_id=device_pk,
-                interface_name=interface_name,
-                ssid=ssid,
-                wifi_client=client_obj,
-                stop_time=None,
-            )
-            active_sessions.append(session_obj.pk)
-
-    # Close open WifiSession
-    WifiSession.objects.filter(device_id=device_pk, stop_time=None,).exclude(
-        pk__in=active_sessions
-    ).update(stop_time=now())
-
-
-@shared_task
-def delete_wifi_clients_and_sessions(days=6 * 30):
-    WifiClient = load_model('monitoring', 'WifiClient')
-    WifiSession = load_model('monitoring', 'WifiSession')
-
-    WifiSession.objects.filter(start_time__lte=(now() - timedelta(days=days))).delete()
-    WifiClient.objects.exclude(
-        mac_address__in=WifiSession.objects.values_list('wifi_client')
-    ).delete()
