+++ openwisp_monitoring/monitoring/tests/__init__.py
@@ -1,11 +1,8 @@
-from copy import deepcopy
 from datetime import timedelta
 
 from django.utils.timezone import now
 from swapper import load_model
 
-from openwisp_controller.config.tests.utils import CreateConfigTemplateMixin
-from openwisp_monitoring.device.utils import manage_short_retention_policy
 from openwisp_users.tests.utils import TestOrganizationMixin
 
 from ...db import timeseries_db
@@ -22,9 +19,6 @@ ten_minutes_ago = start_time - timedelta(minutes=10)
 Chart = load_model('monitoring', 'Chart')
 Metric = load_model('monitoring', 'Metric')
 AlertSettings = load_model('monitoring', 'AlertSettings')
-WifiClient = load_model('monitoring', 'WifiClient')
-WifiSession = load_model('monitoring', 'WifiSession')
-DeviceData = load_model('device_monitoring', 'DeviceData')
 
 default_message = (
     '{notification.actor.name} for device [{notification.target}]'
@@ -241,195 +235,3 @@ class TestMonitoringMixin(TestOrganizationMixin):
         c.full_clean()
         c.save()
         return c
-
-    def _data(self):
-        return {
-            'type': 'DeviceMonitoring',
-            'general': {'local_time': 1589026500, 'uptime': 8003},
-            'resources': {
-                'cpus': 1,
-                'memory': {
-                    'total': 249774080,
-                    'shared': 86016,
-                    'free': 224497664,
-                    'cached': 6774784,
-                    'available': 223397664,
-                    'buffered': 974848,
-                },
-                'load': [0, 0, 0],
-                'disk': [
-                    {
-                        'used_bytes': 18792,
-                        'available_bytes': 233984,
-                        'filesystem': '/dev/root',
-                        'mount_point': '/',
-                        'used_percent': 7,
-                        'size_bytes': 258016,
-                    },
-                    {
-                        'used_bytes': 3872,
-                        'available_bytes': 11916,
-                        'filesystem': '/dev/sda1',
-                        'mount_point': '/boot',
-                        'used_percent': 25,
-                        'size_bytes': 16112,
-                    },
-                ],
-                'swap': {'free': 0, 'total': 0},
-            },
-            'interfaces': [
-                {
-                    'name': 'wlan0',
-                    'type': 'wireless',
-                    'up': True,
-                    'mac': '44:d1:fa:4b:38:43',
-                    'txqueuelen': 1000,
-                    'multicast': True,
-                    'mtu': 1500,
-                    'statistics': {
-                        'rx_bytes': 324,
-                        'tx_bytes': 145,
-                        'collisions': 0,
-                        'multicast': 0,
-                        'rx_dropped': 0,
-                        'tx_dropped': 0,
-                    },
-                    'wireless': {
-                        'frequency': 2437,
-                        'mode': 'access_point',
-                        'signal': -29,
-                        'tx_power': 6,
-                        'channel': 6,
-                        'ssid': 'testnet',
-                        'noise': -95,
-                        'country': 'US',
-                        'clients': [
-                            {
-                                'mac': '00:ee:ad:34:f5:3b',
-                                'wps': False,
-                                'wds': False,
-                                'ht': True,
-                                'preauth': False,
-                                'assoc': True,
-                                'authorized': True,
-                                'vht': False,
-                                'wmm': True,
-                                'aid': 1,
-                                'mfp': False,
-                                'auth': True,
-                                'signature': 'test_signature',
-                            }
-                        ],
-                    },
-                },
-                {
-                    'name': 'wlan1',
-                    'type': 'wireless',
-                    'up': True,
-                    'mac': '44:d1:fa:4b:38:44',
-                    'txqueuelen': 1000,
-                    'multicast': True,
-                    'mtu': 1500,
-                    'statistics': {
-                        'rx_bytes': 2275,
-                        'tx_bytes': 826,
-                        'collisions': 0,
-                        'multicast': 0,
-                        'rx_dropped': 0,
-                        'tx_dropped': 0,
-                    },
-                    'wireless': {
-                        'frequency': 2437,
-                        'mode': 'access_point',
-                        'signal': -29,
-                        'tx_power': 6,
-                        'channel': 6,
-                        'ssid': 'testnet',
-                        'noise': -95,
-                        'country': 'US',
-                        'clients': [
-                            {
-                                'mac': 'b0:e1:7e:30:16:44',
-                                'wps': False,
-                                'wds': False,
-                                'ht': True,
-                                'preauth': False,
-                                'assoc': True,
-                                'authorized': True,
-                                'vht': False,
-                                'wmm': True,
-                                'aid': 1,
-                                'mfp': False,
-                                'auth': True,
-                                'signature': 'test_signature',
-                            },
-                            {
-                                'mac': 'c0:ee:fb:34:f5:4b',
-                                'wps': False,
-                                'wds': False,
-                                'ht': True,
-                                'preauth': False,
-                                'assoc': True,
-                                'authorized': True,
-                                'vht': False,
-                                'wmm': True,
-                                'aid': 1,
-                                'mfp': False,
-                                'auth': True,
-                                'signature': 'test_signature',
-                            },
-                        ],
-                    },
-                },
-            ],
-        }
-
-
-class TestWifiClientSessionMixin(CreateConfigTemplateMixin, TestMonitoringMixin):
-    @classmethod
-    def setUpClass(cls):
-        super().setUpClass()
-        manage_short_retention_policy()
-
-    @property
-    def _sample_data(self):
-        data = deepcopy(self._data())
-        data.pop('resources')
-        return data
-
-    def _create_device_data(self, device=None):
-        device = device or self._create_device()
-        return DeviceData(pk=device.pk)
-
-    def _save_device_data(self, device_data=None, data=None):
-        dd = device_data or self._create_device_data()
-        dd.data = data or self._sample_data
-        dd.save_data()
-        return dd
-
-    def _create_wifi_client(self, **kwargs):
-        options = {
-            'mac_address': '22:33:44:55:66:77',
-            'vendor': '',
-            'ht': True,
-            'vht': True,
-            'wmm': False,
-            'wds': False,
-            'wps': False,
-        }
-        options.update(**kwargs)
-        wifi_client = WifiClient(**options)
-        wifi_client.full_clean()
-        wifi_client.save()
-        return wifi_client
-
-    def _create_wifi_session(self, **kwargs):
-        if 'wifi_client' not in kwargs:
-            kwargs['wifi_client'] = self._create_wifi_client()
-        if 'device' not in kwargs:
-            kwargs['device'] = self._create_device()
-        options = {'ssid': 'Free Public WiFi', 'interface_name': 'wlan0'}
-        options.update(kwargs)
-        wifi_session = WifiSession(**options)
-        wifi_session.full_clean()
-        wifi_session.save()
