+++ tests/openwisp2/sample_device_monitoring/models.py
@@ -5,9 +5,18 @@ from swapper import get_model_name, load_model
 from openwisp_monitoring.device.base.models import (
     AbstractDeviceData,
     AbstractDeviceMonitoring,
+    AbstractWifiClient,
+    AbstractWifiSession,
 )
 
 BaseDevice = load_model('config', 'Device', require_ready=False)
+
+
+class DetailsModel(models.Model):
+    details = models.CharField(max_length=64, default='devicemonitoring')
+
+    class Meta:
+        abstract = True
 
 
 class DeviceData(AbstractDeviceData, BaseDevice):
@@ -18,11 +27,19 @@ class DeviceData(AbstractDeviceData, BaseDevice):
         proxy = True
 
 
-class DeviceMonitoring(AbstractDeviceMonitoring):
-    details = models.CharField(max_length=64, default='devicemonitoring')
-
+class DeviceMonitoring(DetailsModel, AbstractDeviceMonitoring):
     class Meta(AbstractDeviceMonitoring.Meta):
         abstract = False
 
     def __str__(self):
         return self.details
+
+
+class WifiClient(DetailsModel, AbstractWifiClient):
+    class Meta(AbstractWifiClient.Meta):
+        abstract = False
+
+
+class WifiSession(DetailsModel, AbstractWifiSession):
+    class Meta(AbstractWifiSession.Meta):
+        abstract = False
